backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
supabase>=2.3.0
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
pymupdf>=1.24.0
pydantic>=2.9.0
python-multipart>=0.0.9
